Rename test_probes to tests_backend, and probe_{name}.py to {name}.py
