+++ winforms/tests_backend/widgets/label.py
@@ -1,6 +1,6 @@
 import System.Windows.Forms
 
-from .probe_base import SimpleProbe
+from .base import SimpleProbe
 
 
 class LabelProbe(SimpleProbe):
